+++ tests/test_snippets/test_adventures.py
@@ -86,32 +86,6 @@ def collect_snippets(path, filtered_language = None):
 
   return Hedy_snippets
 
-def translate_keywords_in_snippets(snippets):
-    # fill keyword dict for all keyword languages
-    keyword_dict = {}
-    for lang in ALL_KEYWORD_LANGUAGES:
-        keyword_dict[lang] = KEYWORDS.get(lang)
-        for k, v in keyword_dict[lang].items():
-            if type(v) == str and "|" in v:
-                # when we have several options, pick the first one as default
-                keyword_dict[lang][k] = v.split('|')[0]
-    english_keywords = KEYWORDS.get("en")
-
-    # We replace the code snippet placeholders with actual keywords to the code is valid: {print} -> print
-    for snippet in snippets:
-        try:
-            if snippet[1].language in ALL_KEYWORD_LANGUAGES.keys():
-                snippet[1].code = snippet[1].code.format(**keyword_dict[snippet[1].language])
-            else:
-                snippet[1].code = snippet[1].code.format(**english_keywords)
-        except KeyError:
-            print("This following snippet contains an invalid placeholder ...")
-            print(snippet)
-
-    return snippets
-
-
-
 # use this to filter on 1 lang, zh_Hans for Chinese, nb_NO for Norwegian, pt_PT for Portuguese
 # Hedy_snippets = [(s.name, s) for s in collect_snippets(path='../../content/adventures', filtered_language='pt_PT')]
 
@@ -121,7 +95,7 @@ Hedy_snippets = [(s.name, s) for s in collect_snippets(path='../../content/adven
 # if level:
 #     Hedy_snippets = [(name, snippet) for (name, snippet) in Hedy_snippets if snippet.level == level]
 
-Hedy_snippets = translate_keywords_in_snippets(Hedy_snippets)
+Hedy_snippets = HedyTester.translate_keywords_in_snippets(Hedy_snippets)
 
 class TestsAdventurePrograms(unittest.TestCase):
