+++ tests/test_snippets/test_parsons.py
@@ -48,34 +48,8 @@ def collect_snippets(path):
 
     return Hedy_snippets
 
-def translate_keywords_in_snippets(snippets):
-    # fill keyword dict for all keyword languages
-    keyword_dict = {}
-    for lang in ALL_KEYWORD_LANGUAGES:
-        keyword_dict[lang] = YamlFile.for_file(f'../../content/keywords/{lang}.yaml').to_dict()
-        for k, v in keyword_dict[lang].items():
-            if type(v) == str and "|" in v:
-                # when we have several options, pick the first one as default
-                keyword_dict[lang][k] = v.split('|')[0]
-
-    english_keywords = YamlFile.for_file(f'../../content/keywords/en.yaml').to_dict()
-
-    # We replace the code snippet placeholders with actual keywords to the code is valid: {print} -> print
-    for snippet in snippets:
-        try:
-            if snippet[1].language in ALL_KEYWORD_LANGUAGES.keys():
-                snippet[1].code = snippet[1].code.format(**keyword_dict[snippet[1].language])
-            else:
-                snippet[1].code = snippet[1].code.format(**english_keywords)
-        except KeyError:
-            print("This following snippet contains an invalid placeholder ...")
-            print(snippet)
-
-    return snippets
-
-
 Hedy_snippets = [(s.name, s) for s in collect_snippets(path='../../content/parsons')]
-Hedy_snippets = translate_keywords_in_snippets(Hedy_snippets)
+Hedy_snippets = HedyTester.translate_keywords_in_snippets(Hedy_snippets)
 
 
 class TestsParsonsPrograms(unittest.TestCase):
